clk/qcom: apq8016: add support for USB_HS clocks
The newer "register map for simple gate clocks" support added for qcom
clocks is used. As a result gcc_apq8016 now has a mixture of the old and
new styles. I didn't (and still don't!) feel comfortable enough in this
area to update the existing code.

Some groundwork was also laid to dump the state of these gate clocks
with the recently-added qcom dump callbacks.

Signed-off-by: Sam Day <me@samcday.com>