usb: gadget: ci_udc: Fix crash if usb_ep_disable() is called on disabled EP
f_acm calls usb_ep_disable() during initialization for an EP that wasn't
enabled before. This works fine on the DWC3 UDC driver, but on ci_udc this
crashes at the moment, because we don't have any checks for that.

Make sure the ep->desc is not NULL to avoid that crash.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>