video: Add minimal driver for Qualcomm MDP5
The Qualcomm Mobile Display Processor (MDP) is part of the Multimedia
Display Subsystem (MDSS) IP block. It's quite a complex beast made up of
several hundred registers that control a backend composition engine that
handles scanning framebuffer memory out to the DSI PHY, with lots of
bells and whistles.

A proper U-Boot driver that can configure this system would be great.
This newly introduced driver is surely not that. Instead, it's a simple
implementation that does an any% speedrun to discovering an existing MDP
"pipe" that was configured by a previous bootloader. It's intended for
use on devices with a vendor bootloader that has already done the
needful configuration.