mach-snapdragon: pass fdt to qcom_parse_memory
With the recent shuffling of board_fdt_blob_setup, the semantics of when
the gd->fdt_blob global is set have changed. Rather than relying on this
subtle behaviour, explicitly pass the FDT that should be consulted for a
/memory node.

Using the OF API is typically preferable because it's easier to read,
but using the lower level fdt_ methods instead here doesn't add too much
complexity, I think.

Also a minor tweak to make sure we're looking at the fdt blob passed to
board_fdt_blob_setup, which removes the last of the references to global
data in this function.

Signed-off-by: Sam Day <me@samcday.com>