mach-snapdragon: use PSCI sysreset driver
Drop the `board_reset` function from snapdragon's board code, and
instead use the standard PSCI sysreset driver.

Other than dropping a little bit of extraneous code, the main goal here
is to pave the way for proper sysreset support on the msm8916 platform.
A bunch of these devices shipped with a TZ firmware that seems to no-op
the reset+poweroff PSCI functions, and instead the PSHOLD driver should
be used.

It's for this reason that the PSCI sysreset driver is deliberately *not*
added as a dependency of ARCH_SNAPDRAGON, and instead included in
qcom_defconfig.

Signed-off-by: Sam Day <me@samcday.com>