button: qcom-pmic: introduce exported setup method
The qcom_pwrkey driver is currently binding to both the individual
pwrkey/resin nodes, as well as the parent pon node.

In the next commit, a new glue driver will be introduced for pm8916_pon,
which will continue binding these buttons, as well as additional
drivers.

button_qcom_pmic_bind was renamed to button_qcom_pmic_setup and exported
in a shared header, which the new pm8916_pon driver will call.