sysreset: qcom-pshold: PM8916 PON reset type support
Before de-asserting PS_HOLD, PM8916 PMIC can be informed of what kind of
reset should take place.

This might seem a little hacky/weird, but it's a byproduct of how the DT
nodes are structured, and the way u-boot handles sysreset.

The qcom,pshold node does not belong to the PMIC hierarchy. If U-Boot's
sysreset infra was a little more flexible, we might be able to model the
qcom,pm8916-pon driver as UCLASS_SYSRESET in order to catch the sysreset
type and set the appropriate bits in the PMIC, before chaining to the
pshold driver to de-assert PS_HOLD (which actually triggers the reset
type that was written in the previous step).

Signed-off-by: Sam Day <me@samcday.com>