qcom_defconfig: enable poweroff command
With the newly introduced PM8916 PON driver and support in PSHOLD
driver, the poweroff command now works on MSM8916 devices.

This should also work on any more recent qcom SoC with a functional PSCI
implementation in TZ. I confirmed that is the case on my google-sargo
and oneplus-fajita, at least.

Signed-off-by: Sam Day <me@samcday.com>