MSM8916: Bring secondary cores online with spin-table
This patch series depends on the "mach-snapdragon: handle platforms
without PSCI support" patch that was submitted earlier today. That first
patch lays some groundwork to handle detecting when a device is booting
without PSCI and deletes the /psci node in that case.

This series builds on top of that detection by rewriting all CPUs with
enable-method="psci" to "spin-table".

Once this is done, the existing spin-table support in U-Boot handles
further patching the nodes with the appropriate cpu-release-addr and
reserving the memory with the spin-table instructions.

Interestingly, this spin-table code doesn't seem to be in active use
anywhere, and it had bitrotted a little. But after blowing the dust off
it and fixing a couple of missing #includes, it works great.

I extended the spin-table support a little to make it easier for
consumers to plug in a method that is called for each CPU that needs to
be brought online. This saves the qcom board code from needing to hook
into ft_board_setup or something that runs before booting the OS.

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

Signed-off-by: Sam Day <me@samcday.com>
---
Changes in v3:
- Fix up attribution/copyright in scm/smp code
- qcom_boot_cortex_a53: Issue sync barriers when poking APCS registers
- qcom_scm_set_boot_addr_mc: Delay qcom_scm_set_boot_addr_mc as late as possible
- qcom_scm_set_boot_addr_mc: Remove unnecessary msm8916 compatible check
- Link to v2: https://lore.kernel.org/r/20260602-msm8916-smp-support-v2-0-792e53dac374@samcday.com

Changes in v2:
- Reworked the SCM impl to be a proper firmware driver, and streamlined
  the API so callers aren't forced to deal with calling conventions
  (thanks Casey).
- Link to v1: https://lore.kernel.org/r/20250128-msm8916-smp-support-v1-0-39587f18cf0c@samcday.com

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20250127-msm8916-smp-support-5f5e7b49d07e",
    "prefixes": [],
    "prerequisites": [],
    "history": {
      "v1": [
        "20250128-msm8916-smp-support-v1-0-39587f18cf0c@samcday.com"
      ],
      "v2": [
        "20260602-msm8916-smp-support-v2-0-792e53dac374@samcday.com"
      ]
    }
  }
}