armv8: spin_table: don't overwrite cpu-release-addr
It might be the case that a previous bootloader has already prepared the
spin-table memory and patched the CPU nodes. In this case we should not
try to boot these CPUs or overwrite the existing cpu-release-addr.

In addition, if we didn't find any CPUs to boot, then there's no need to
add a reservation for the spin-table memory.

Signed-off-by: Sam Day <me@samcday.com>